hercdectables: Glossaries that Explain the Census API's Decennial Group Tables

A data package that contains glossaries for US Census groups. A glossary is a mapping from variable names to a collection of explicit values that the variables encode. For example, a variable in the Census database might report a count of the number of people from one demographic group living in one type of housing and primarily speaking one language. The glossary for that group would include columns of demographic group, housing type, and primary language. These glossaries explain the Census API's Decennial group tables.

See also

Author

Maintainer: Ben Taft ben.taft@career2cradle.org

Other contributors:

Higher Expectations for Racine County [copyright holder]